+++ src/main/java/jscover/server/ConfigurationForServer.java
@@ -351,6 +351,8 @@ import org.mozilla.javascript.Context;
 import java.io.File;
 import java.util.HashSet;
 import java.util.Set;
+import java.util.regex.Pattern;
+import java.util.regex.PatternSyntaxException;
 
 public class ConfigurationForServer extends Configuration {
     public static final String HELP_PREFIX1 = Main.HELP_PREFIX1;
@@ -359,6 +361,7 @@ public class ConfigurationForServer extends Configuration {
     public static final String PORT_PREFIX = "--port=";
     public static final String REPORT_DIR_PREFIX = "--report-dir=";
     public static final String NO_INSTRUMENT_PREFIX = "--no-instrument=";
+    public static final String NO_INSTRUMENT_REG_PREFIX = "--no-instrument-reg=";
     public static final String JS_VERSION_PREFIX = "--js-version=";
     public static final String PROXY_PREFIX = "--proxy";
     public static final String INCLUDE_UNLOADED_JS_PREFIX = "--include-unloaded-js";
@@ -370,6 +373,7 @@ public class ConfigurationForServer extends Configuration {
     private File documentRoot = new File(System.getProperty("user.dir"));
     private Integer port = 8080;
     private final Set<String> noInstruments = new HashSet<String>();
+    private final Set<Pattern> noInstrumentRegs = new HashSet<Pattern>();
     private File reportDir = new File(System.getProperty("user.dir"));
     private int JSVersion = Context.VERSION_1_5;
     private boolean proxy;
@@ -406,10 +410,12 @@ public class ConfigurationForServer extends Configuration {
     }
 
     public boolean skipInstrumentation(String uri) {
-        for (String noInstrument : noInstruments) {
+        for (String noInstrument : noInstruments)
             if (uri.startsWith(noInstrument))
                 return true;
-        }
+        for (Pattern noInstrumentReg : noInstrumentRegs)
+            if (noInstrumentReg.matcher(uri).matches())
+                return true;
         return false;
     }
 
@@ -436,6 +442,18 @@ public class ConfigurationForServer extends Configuration {
                 if (uri.startsWith("/"))
                     uri = uri.substring(1);
                 configuration.noInstruments.add(uri);
+            } else if (arg.startsWith(NO_INSTRUMENT_REG_PREFIX)) {
+                String patternString = arg.substring(NO_INSTRUMENT_REG_PREFIX.length());
+                if (patternString.startsWith("/"))
+                    patternString = patternString.substring(1);
+                try {
+                    Pattern pattern = Pattern.compile(patternString);
+                    configuration.noInstrumentRegs.add(pattern);
+                } catch(PatternSyntaxException e) {
+                    e.printStackTrace(System.err);
+                    configuration.showHelp = true;
+                    configuration.invalid = true;
+                }
             } else if (arg.startsWith(JS_VERSION_PREFIX)) {
                 configuration.JSVersion = (int)(Float.valueOf(arg.substring(JS_VERSION_PREFIX.length()))*100);
             } else if (arg.equals(BRANCH_PREFIX)) {