+++ History.md
@@ -1,5 +1,6 @@
 0.2.7 / 2013-??
 ==================
+  * Add regular expression based path switch support (https://github.com/tntim96/JSCover/issues/57)
   * Add HTTP Server support for XML MIME type (https://github.com/tntim96/JSCover/issues/56)
 
 0.2.6 / 2013-03-11